+++ src/index.tsx
@@ -18,6 +18,7 @@ export { default as BaseSection, SectionContentSize } from 'layout/section/baseS
 export { default as HeroSection } from 'layout/section/heroSection'
 export { default as HighlightSection } from 'layout/section/highlightSection'
 export { default as TabsSection } from 'layout/section/tabsSection'
+export { default as TripsSection } from 'layout/section/tripsSection'
 export { default as ItemAction } from 'itemAction'
 export { default as ItemCheckbox, ItemCheckboxStatus } from 'itemCheckbox'
 export { default as ItemChoice, ItemChoiceStyle, ItemChoiceStatus } from 'itemChoice'
@@ -53,7 +54,6 @@ export { default as TimePicker } from 'timePicker'
 export { default as Title } from 'title'
 export { default as ToggleButton, ToggleButtonStatus } from 'toggleButton'
 export { default as TopBar } from 'topBar'
-export { default as TopTripCardList } from 'topTripCardList'
 export { default as Transitions, AnimationType } from 'transitions'
 export { default as TripCard } from 'tripCard'
 export { default as UneditableTextField } from 'uneditableTextField'
