+++ stories/index.tsx
@@ -5,6 +5,7 @@ import '../src/layout/section/baseSection/story'
 import '../src/layout/section/heroSection/story'
 import '../src/layout/section/tabsSection/story'
 import '../src/layout/section/highlightSection/story'
+import '../src/layout/section/tripsSection/story'
 
 // Pages
 import '../src/pages/messaging/inbox.story'
@@ -64,7 +65,6 @@ import '../src/timePicker/story'
 import '../src/title/story'
 import '../src/toggleButton/story'
 import '../src/topBar/story'
-import '../src/topTripCardList/story'
 import '../src/tripCard/story'
 import '../src/uneditableTextField/story'
 import '../src/why/story'
